@@ -5,7 +5,6 @@
   "angular-cli": {},
   "scripts": {
     "start": "ng serve",
-    "postinstall": "typings install",
     "lint": "tslint \"src/**/*.ts\"",
     "test": "ng test",
     "pree2e": "webdriver-manager update",
@@ -21,30 +20,28 @@
     "@angular/platform-browser": "2.0.0-rc.5",
     "@angular/platform-browser-dynamic": "2.0.0-rc.5",
     "@angular/router": "3.0.0-rc.1",
-    "@angular/router-deprecated": "2.0.0-rc.2",
     "bootstrap-css-only": "^3.3.6",
-    "es6-shim": "0.35.1",
+    "core-js": "^2.4.0",
     "font-awesome": "^4.6.3",
     "leaflet": "^0.7.7",
-    "reflect-metadata": "0.1.3",
-    "rxjs": "5.0.0-beta.6",
-    "systemjs": "0.19.26",
+    "rxjs": "5.0.0-beta.11",
+    "ts-helpers": "^1.1.1",
     "zone.js": "0.6.12"
   },
   "devDependencies": {
-    "@angular/tsc-wrapped": "^0.2.2",
-    "angular-cli": "^1.0.0-beta.10",
-    "codelyzer": "0.0.20",
-    "ember-cli-inject-live-reload": "1.4.0",
+    "@types/jasmine": "^2.2.30",
+    "@types/leaflet": "^1.0.31",
+    "angular-cli": "1.0.0-beta.11-webpack.8",
+    "codelyzer": "~0.0.26",
     "jasmine-core": "2.4.1",
     "jasmine-spec-reporter": "2.5.0",
     "karma": "0.13.22",
     "karma-chrome-launcher": "0.2.3",
     "karma-jasmine": "0.3.8",
-    "protractor": "3.3.0",
-    "ts-node": "0.5.5",
-    "tslint": "3.11.0",
-    "typescript": "^1.9.0-dev",
-    "typings": "1.3.1"
+    "karma-remap-istanbul": "^0.2.1",
+    "protractor": "4.0.3",
+    "ts-node": "1.2.1",
+    "tslint": "3.13.0",
+    "typescript": "^2.0.0"
   }
 }
