@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Http, Response } from '@angular/http';
 
-import {Observable} from "rxjs/Rx";
+import {Observable} from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
 import 'rxjs/add/observable/of';
@@ -48,8 +48,8 @@ export class PlaygroundService {
               'lat': openPlayground.geometry.coordinates[0][1],
               'lng': openPlayground.geometry.coordinates[0][0]
             }
-          }
-        })
+          };
+        });
       })
       .publishLast()
       .refCount()
