@@ -1,8 +1,7 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, Input } from '@angular/core';
 
 import { Playground } from '../shared';
 @Component({
-  moduleId: module.id,
   selector: 'app-footer',
   templateUrl: 'footer.component.html',
   styleUrls: ['footer.component.css']
